Refactor: Replace obsolete OrdinatedPageDto with IOrdinatedPageDto and update related methods for improved paging functionality
