+++ Carbon.Common/IOrdinatedPageDto.cs
@@ -1,0 +1,12 @@
+using System;
+
+namespace Carbon.Common
+{
+    /// <summary>
+    /// Represents the interface for the ordinated paged data transfer object.
+    /// </summary>
+    public interface IOrdinatedPageDto :IOrderableDto, IPageableDto
+    {
+
+    }
+}
